moving to own file, adding WrappedWaitGroup, changing loop condition
