+++ ipld/get.go
@@ -2,17 +2,12 @@ package ipld
 
 import (
 	"context"
-	"sort"
-	"sync"
 
 	"github.com/ipfs/go-blockservice"
 	"github.com/ipfs/go-cid"
 	ipld "github.com/ipfs/go-ipld-format"
 
 	"github.com/celestiaorg/celestia-node/ipld/plugin"
-	"github.com/celestiaorg/nmt"
-
-	"github.com/celestiaorg/nmt/namespace"
 )
 
 // GetShare fetches and returns the data for leaf `leafIndex` of root `rootCid`.
@@ -126,126 +121,6 @@ func GetProof(
 
 	// recursively walk down through selected children
 	return GetProof(ctx, bGetter, root, proof, leaf, total)
-}
-
-// GetSharesByNamespace returns all the shares from the given root
-// with the given namespace.ID.
-func GetSharesByNamespace(
-	ctx context.Context,
-	bGetter blockservice.BlockGetter,
-	root cid.Cid,
-	nID namespace.ID,
-) ([]Share, error) {
-	leaves := GetLeavesByNamespace(ctx, bGetter, root, nID)
-	// if err != nil {
-	// 	return nil, err
-	// }
-
-	shares := make([]Share, len(leaves))
-	for i, leaf := range leaves {
-		shares[i] = leafToShare(leaf)
-	}
-
-	return shares, nil
-}
-
-// GetLeavesByNamespace returns all the leaves from the given root with the given namespace.ID.
-// If nothing is found it returns data as nil.
-func GetLeavesByNamespace(
-	ctx context.Context,
-	bGetter blockservice.BlockGetter,
-	root cid.Cid,
-	nID namespace.ID,
-) []ipld.Node {
-	type job struct {
-		id  cid.Cid
-		pos int
-	}
-
-	// TODO: There has to be a more elegant solution
-	// bookkeeping is needed to be able to sort the leaves after the walk
-	type result struct {
-		node ipld.Node
-		pos  int
-	}
-
-	// TODO: Should this be NumWorkersLimit?
-	// we don't know the amount of shares in the namespace, so we cannot preallocate properly
-	jobs := make(chan *job, NumWorkersLimit)
-	jobs <- &job{id: root}
-
-	// the wg counter cannot be preallocated either, it is incremented with each job
-	wg := sync.WaitGroup{}
-	wg.Add(1)
-
-	var leaves []result
-	mu := &sync.Mutex{}
-
-	// once the wait group is done, we can close the jobs channel to break the following loop
-	go func() {
-		wg.Wait()
-		close(jobs)
-	}()
-
-	for j := range jobs {
-		j := j
-		// TODO: This is using the pool from get_shares.go, is this okay?
-		pool.Submit(func() {
-			defer wg.Done()
-			err := SanityCheckNID(nID)
-			if err != nil {
-				return
-			}
-
-			rootH := plugin.NamespacedSha256FromCID(j.id)
-			if nID.Less(nmt.MinNamespace(rootH, nID.Size())) || !nID.LessOrEqual(nmt.MaxNamespace(rootH, nID.Size())) {
-				return
-			}
-
-			nd, err := plugin.GetNode(ctx, bGetter, j.id)
-			if err != nil {
-				return
-			}
-
-			lnks := nd.Links()
-			// wg counter should be incremented before adding new jobs
-			if len(lnks) > 1 {
-				wg.Add(len(lnks))
-			}
-
-			if len(lnks) == 1 {
-				mu.Lock()
-				leaves = append(leaves, result{nd, j.pos})
-				mu.Unlock()
-				return
-			}
-
-			for i, lnk := range lnks {
-				select {
-				case jobs <- &job{
-					id:  lnk.Cid,
-					pos: j.pos*2 + i,
-				}:
-				case <-ctx.Done():
-					return
-				}
-			}
-		})
-	}
-
-	if len(leaves) > 0 {
-		sort.SliceStable(leaves, func(i, j int) bool {
-			return leaves[i].pos < leaves[j].pos
-		})
-
-		output := make([]ipld.Node, len(leaves))
-		for i, leaf := range leaves {
-			output[i] = leaf.node
-		}
-		return output
-	}
-
-	return nil
 }
 
 // leafToShare converts an NMT leaf into a Share.
